refactor: 분기 처리를 위한 에러 추가및 리팩토링
