+++ app/src/main/java/woowacourse/shopping/data/repository/RecentlyProductRepositoryImpl.kt
@@ -5,7 +5,7 @@ import woowacourse.shopping.data.db.recently.RecentlyProductDatabase
 import woowacourse.shopping.data.model.RecentlyProductEntity
 import woowacourse.shopping.domain.model.RecentlyProduct
 import woowacourse.shopping.domain.repository.RecentlyProductRepository
-import woowacourse.shopping.utils.NoSuchDataException
+import woowacourse.shopping.utils.exception.NoSuchDataException
 import kotlin.concurrent.thread
 
 class RecentlyProductRepositoryImpl(context: Context) : RecentlyProductRepository {
