+++ app/src/main/java/woowacourse/shopping/utils/exception/DataIdException.kt
@@ -1,0 +1,7 @@
+package woowacourse.shopping.utils.exception
+
+class DataIdException(message: String = DATA_ID_MESSAGE): Exception(message) {
+    companion object {
+        private const val DATA_ID_MESSAGE = "No such data ID"
+    }
+}
