+++ app/src/main/java/woowacourse/shopping/utils/exception/NoSuchDataException.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.utils
+package woowacourse.shopping.utils.exception
 
 class NoSuchDataException(message: String = DEFAULT_MESSAGE) : Exception(message) {
     companion object {
