+++ app/src/main/java/woowacourse/shopping/view/detail/ProductDetailFragment.kt
@@ -13,7 +13,7 @@ import woowacourse.shopping.data.repository.ShoppingCartRepositoryImpl
 import woowacourse.shopping.databinding.FragmentProductDetailBinding
 import woowacourse.shopping.domain.model.Product
 import woowacourse.shopping.domain.model.RecentlyProduct
-import woowacourse.shopping.utils.NoSuchDataException
+import woowacourse.shopping.utils.exception.NoSuchDataException
 import woowacourse.shopping.utils.ShoppingUtils.makeToast
 import woowacourse.shopping.view.MainActivityListener
 import woowacourse.shopping.view.ViewModelFactory
