+++ app/src/main/java/woowacourse/shopping/view/products/ProductListViewModel.kt
@@ -13,7 +13,7 @@ import woowacourse.shopping.domain.repository.ProductRepository
 import woowacourse.shopping.domain.repository.RecentlyProductRepository
 import woowacourse.shopping.domain.repository.ShoppingCartRepository
 import woowacourse.shopping.utils.MutableSingleLiveData
-import woowacourse.shopping.utils.NoSuchDataException
+import woowacourse.shopping.utils.exception.NoSuchDataException
 import woowacourse.shopping.utils.SingleLiveData
 import woowacourse.shopping.view.cartcounter.ChangeCartItemResultState
